Moved V3 example parsing into OpenApiResponseDeserializer
